requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
distro==1.9.0
fastapi==0.115.11
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
loguru==0.7.3
magic-filter==1.0.12
multidict==6.1.0
netaddr==1.3.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
yarl==1.18.3
yookassa==3.5.0
